minor perf improvements

+++ src/ipc.rs
@@ -1,7 +1,8 @@
 pub enum MainCommands {
     Recalc,
 }
+
 pub enum ClientCommands {
     Redraw(Vec<u32>),
-    RedrawPixel((usize, u32)),
+    RedrawPixel((u32, u32)),
 }
